+++ Tests/SwiftStdlibTests/ArrayExtensionsTests.swift
@@ -281,7 +281,7 @@ final class ArrayExtensionsTests: XCTestCase {
     func testGroupBy() {
         let array: [String] = ["james", "irving", "jordan", "jonshon", "iverson"]
         let grouped = array.groupByKey { element -> String in
-            return String(element.characters.first ?? Character(""))
+            return String(element.first!)
         }
         XCTAssertEqual(grouped["j"] ?? [], [ "james", "jordan", "jonshon" ])
         XCTAssertEqual(grouped["i"] ?? [], [ "irving", "iverson" ])